requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-RethinkDB==0.2
Flask-Scrypt==0.1.3.6
Flask-WTF==0.12

